requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1607439979954/work
aiohttp @ file:///C:/ci/aiohttp_1602530358780/work
albumentations==1.0.3
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828549974/work
arrow @ file:///C:/ci/arrow_1599270424454/work
astroid==2.4.2
astropy==4.0.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attr==0.3.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall==0.2.0
binaryornot==0.4.4
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
blinker==1.4
blis @ file:///C:/ci/cython-blis_1594248830923/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1607706694405/work
catalogue==1.0.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1606255207413/work
chardet @ file:///C:/ci/chardet_1605303259695/work
click==7.1.2
colorama==0.4.3
cookiecutter==1.7.2
coremltools==4.1
cryptography @ file:///C:/ci/cryptography_1607639129468/work
cycler==0.10.0
cymem @ file:///C:/ci/cymem_1605062363169/work
decorator==4.4.2
defusedxml==0.6.0
efficientnet-pytorch==0.6.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
Flask==1.1.2
future==0.18.2
gast==0.3.3
google-auth @ file:///tmp/build/80754af9/google-auth_1607969906642/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1597406403308/work
gunicorn==20.0.4
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.9.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1616837008098/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipython-sql @ file:///D:/bld/ipython-sql_1588134619228/work
isort==5.5.2
itsdangerous==1.1.0
jedi @ file:///C:/ci/jedi_1608902611503/work
Jinja2==2.11.2
jinja2-time==0.2.0
joblib @ file:///tmp/build/80754af9/joblib_1607970656719/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///C:/ci/jupyter_core_1606130908768/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1603996595157/work
lazy-object-proxy==1.4.3
Markdown @ file:///C:/ci/markdown_1605111187600/work
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib @ file:///D:/bld/matplotlib-base_1595215719671/work
mccabe==0.6.1
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mpmath==1.2.1
multidict @ file:///C:/ci/multidict_1600456486794/work
munch==2.5.0
murmurhash @ file:///C:/ci/murmurhash_1607438161064/work
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601914921407/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610664066866/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
networkx==2.6.2
nltk @ file:///tmp/build/80754af9/nltk_1592496090529/work
notebook @ file:///C:/ci/notebook_1596837179121/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
oauthlib==3.1.0
olefile==0.46
onnx==1.9.0
onnx-coreml==1.3
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1607971725249/work
pandas @ file:///C:/ci/pandas_1598371525254/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso==0.7.0
pickle5 @ file:///D:/bld/pickle5_1592859470301/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1609786872067/work
plac @ file:///C:/ci/plac_1594257364512/work
poyo==0.5.0
preshed==3.0.2
pretrainedmodels==0.7.4
prettytable==0.7.2
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
python-slugify @ file:///tmp/build/80754af9/python-slugify_1600462544980/work
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
regex @ file:///C:/ci/regex_1606691216163/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1610483308194/work
scikit-image==0.18.3
scikit-learn @ file:///C:/ci/scikit-learn_1598376983131/work
scipy @ file:///C:/ci/scipy_1597686737426/work
seaborn @ file:///tmp/build/80754af9/seaborn_1600553570093/work
segmentation-models-pytorch==0.2.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
six==1.15.0
spacy @ file:///D:/bld/spacy_1608283203247/work
SQLAlchemy @ file:///C:/ci/sqlalchemy_1608243783929/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
srsly @ file:///C:/ci/srsly_1607530581965/work
sympy==1.8
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
thinc @ file:///C:/ci/thinc_1607710279177/work
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2021.8.30
timm==0.4.12
toml==0.10.1
torch==1.5.1
torchvision==0.6.1
tornado @ file:///C:/ci/tornado_1606935947090/work
tqdm @ file:///tmp/build/80754af9/tqdm_1609788246169/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typed-ast==1.4.1
typing==3.7.4.3
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1602702424206/work
Unidecode @ file:///tmp/build/80754af9/unidecode_1608579687475/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1606938623459/work
waitress==1.4.4
wasabi @ file:///tmp/build/80754af9/wasabi_1599244993840/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.12.1
yarl @ file:///C:/ci/yarl_1606940114600/work
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
